academico/requirements.txt
Django==3.1.2
gunicorn == 20.0.4 
psycopg2-binary==2.8.5
django-ckeditor
django-ckeditor-widget
django-phone-field
djangorestframework
djangorestframework-simplejwt 
django-model-utils
django-jsonfield